Moved qt stuff out of rj_common/Utils.hpp to improve compile times, removed unused <rj_common/Utils.hpp>, precompile Point.hpp.
