+++ rj_common/include/rj_common/qt_utils.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <QtCore/QDir>
+
+/**
+ * @return Absolute path to the 'run' directory.
+ */
+QDir ApplicationRunDirectory();
