+++ soccer/include/rj_vision_filter/ball/KalmanBall.hpp
@@ -3,7 +3,6 @@
 #include <Configuration.hpp>
 #include <Geometry2d/Point.hpp>
 #include <boost/circular_buffer.hpp>
-#include <rj_common/Utils.hpp>
 #include <rj_vision_filter/ball/CameraBall.hpp>
 #include <rj_vision_filter/filter/KalmanFilter2D.hpp>
