+++ soccer/include/rj_vision_filter/ball/WorldBall.hpp
@@ -3,7 +3,6 @@
 #include <Configuration.hpp>
 #include <Geometry2d/Point.hpp>
 #include <list>
-#include <rj_common/Utils.hpp>
 #include <rj_vision_filter/ball/KalmanBall.hpp>
 
 class KalmanBall;
