+++ soccer/include/rj_vision_filter/camera/CameraFrame.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include <rj_common/Utils.hpp>
 #include <rj_msgs/msg/detection_frame.hpp>
 #include <rj_vision_filter/ball/CameraBall.hpp>
 #include <rj_vision_filter/robot/CameraRobot.hpp>
