+++ soccer/include/rj_vision_filter/kick/KickEvent.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <deque>
-#include <rj_common/Utils.hpp>
 #include <rj_vision_filter/ball/WorldBall.hpp>
 #include <rj_vision_filter/kick/VisionState.hpp>
 #include <rj_vision_filter/robot/WorldRobot.hpp>
