+++ soccer/include/rj_vision_filter/kick/detector/SlowKickDetector.hpp
@@ -3,7 +3,6 @@
 #include <Configuration.hpp>
 #include <Geometry2d/Point.hpp>
 #include <deque>
-#include <rj_common/Utils.hpp>
 #include <rj_vision_filter/ball/WorldBall.hpp>
 #include <rj_vision_filter/kick/KickEvent.hpp>
 #include <rj_vision_filter/kick/VisionState.hpp>
