+++ soccer/src/rj_vision_filter/ball/BallBounce.cpp
@@ -1,6 +1,6 @@
 #include <algorithm>
 #include <cmath>
-#include <rj_common/Utils.hpp>
+#include <Geometry2d/Line.hpp>
 #include <rj_constants/constants.hpp>
 #include <rj_vision_filter/ball/BallBounce.hpp>
 #include <rj_vision_filter/util/VisionFilterConfig.hpp>
