+++ soccer/src/soccer/SystemState.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <GameState.hpp>
-#include <rj_common/Utils.hpp>
 #include <rj_constants/constants.hpp>
 #include <vector>
