+++ soccer/src/soccer/joystick/SDLJoystickNode.cpp
@@ -2,7 +2,6 @@
 
 #include <ament_index_cpp/get_package_share_directory.hpp>
 #include <iostream>
-#include <rj_common/Utils.hpp>
 
 namespace joystick {
 SDLJoystickNode::SDLJoystickNode(Context* context) : context_{context} {
