+++ soccer/src/soccer/planning/Instant.hpp
@@ -3,7 +3,6 @@
 #include <Geometry2d/Point.hpp>
 #include <Geometry2d/Pose.hpp>
 #include <optional>
-#include <rj_common/Utils.hpp>
 #include <rj_common/time.hpp>
 
 #include "DebugDrawer.hpp"
