+++ soccer/src/soccer/planning/Trajectory.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <Geometry2d/Pose.hpp>
-#include <rj_common/Utils.hpp>
 #include <rj_common/time.hpp>
 
 #include "DebugDrawer.hpp"
