+++ soccer/src/soccer/planning/planner/CollectPlanner.cpp
@@ -1,6 +1,5 @@
 #include "CollectPlanner.hpp"
 
-#include <rj_common/Utils.hpp>
 #include <rj_constants/constants.hpp>
 
 #include "Configuration.hpp"
