+++ soccer/src/soccer/planning/planner/MotionCommand.hpp
@@ -3,7 +3,6 @@
 #include <Geometry2d/Point.hpp>
 #include <Geometry2d/Pose.hpp>
 #include <WorldState.hpp>
-#include <rj_common/Utils.hpp>
 #include <variant>
 
 #include "planning/Instant.hpp"
